Make call to action more noticeable.
The call to action (CTA) should be up-front and personal to get as many
conversions as possible. Having the CTA to the side breaks flow. Usually
if the CTA in a template is to the side like that, you're intended to
place an image to the side to use up the negative space.

Additionally, using line break elements is discouraged where CSS can be
used instead.

Sadly, I couldn't come up with an easy "silver bullet" fix, so I had to
modify the CTA directly in the theme's CSS.